logic error corrected
in old code, port in CMDARGS is a unsigned short int
since checkPort returns -1 to port if port provided is not valid, we cannot detect a such a problem.
now port in CMDARGS is just an int.

+++ source/misc.h
@@ -11,7 +11,7 @@
 struct CMDARGS {
     int mode; // rx = 1, tx = 2
     std::string ip; // ip stored as a string
-    unsigned short int port; // port number stored as a unsigned short int
+    int port; // port number stored as a unsigned short int
     std::string filePath; // path to the file, or directory if -d is indicated
     bool dir; // indicate of filePath points to a folder
 };
